functions/main.py
# Welcome to Cloud Functions for Firebase for Python!
# To get started, simply uncomment the below code or create your own.
# Deploy with `firebase deploy`

import os
from firebase_functions import https_fn
from firebase_admin import initialize_app
from google.cloud import vision
from google.cloud import translate_v2 as translate
import functions_framework
from flask import jsonify, Request
import json

# Initialize Firebase Admin
initialize_app()

# Initialize Google Cloud clients
vision_client = vision.ImageAnnotatorClient()
translate_client = translate.Client()

@https_fn.on_request()
def search_drugs(req: Request):
    """Search for drugs by name."""
    if req.method != 'GET':
        return https_fn.Response('Method not allowed', status=405)
    
    name = req.args.get('name')
    if not name:
        return https_fn.Response('Name parameter is required', status=400)
    
    # TODO: Implement drug search logic
    # This is a placeholder. You'll need to implement the actual search logic
    drugs = [
        {
            "name": name,
            "confidence": 0.95,
            "alternatives": []
        }
    ]
    
    return https_fn.Response(
        json.dumps(drugs),
        status=200,
        content_type='application/json'
    )

@https_fn.on_request()
def process_prescription(req: Request):
    """Process prescription image using Google Vision API."""
    if req.method != 'POST':
        return https_fn.Response('Method not allowed', status=405)
    
    if not req.files or 'file' not in req.files:
        return https_fn.Response('No file provided', status=400)
    
    file = req.files['file']
    content = file.read()
    
    # Perform OCR using Google Vision API
    image = vision.Image(content=content)
    response = vision_client.text_detection(image=image)
    texts = response.text_annotations
    
    if not texts:
        return https_fn.Response('No text found in image', status=400)
    
    extracted_text = texts[0].description
    
    # TODO: Process the extracted text to identify medicines
    # This is a placeholder. You'll need to implement the actual processing logic
    result = {
        "text": extracted_text,
        "medicines": []
    }
    
    return https_fn.Response(
        json.dumps(result),
        status=200,
        content_type='application/json'
    )

@https_fn.on_request()
def search_stores(req: Request):
    """Search for medical stores."""
    if req.method != 'GET':
        return https_fn.Response('Method not allowed', status=405)
    
    latitude = req.args.get('latitude')
    longitude = req.args.get('longitude')
    
    if not latitude or not longitude:
        return https_fn.Response('Latitude and longitude are required', status=400)
    
    # TODO: Implement store search logic
    # This is a placeholder. You'll need to implement the actual search logic
    stores = [
        {
            "name": "Sample Medical Store",
            "address": "123 Healthcare Street",
            "distance": "1.2 km",
            "rating": 4.5
        }
    ]
    
    return https_fn.Response(
        json.dumps(stores),
        status=200,
        content_type='application/json'
    )